Split out generate-feed
